+++ src/machines/java/com/enderio/machines/common/init/MachineBlocks.java
@@ -123,15 +123,11 @@ public class MachineBlocks {
     public static final BlockEntry<ProgressMachineBlock> PAINTING_MACHINE = standardMachine("painting_machine", () -> MachineBlockEntities.PAINTING_MACHINE)
         .register();
 
-    public static final BlockEntry<MachineBlock> CREATIVE_POWER = REGISTRATE
-        .block("creative_power", props -> new MachineBlock(props, MachineBlockEntities.CREATIVE_POWER))
-        .item()
-        .tab(EIOCreativeTabs.MACHINES)
-        .build()
+    public static final BlockEntry<ProgressMachineBlock> WIRED_CHARGER = standardMachine("wired_charger", () -> MachineBlockEntities.WIRED_CHARGER)
         .register();
 
-    public static final BlockEntry<MachineBlock> WIRED_CHARGER = REGISTRATE
-        .block("wired_charger", props -> new MachineBlock(props, MachineBlockEntities.WIRED_CHARGER))
+    public static final BlockEntry<MachineBlock> CREATIVE_POWER = REGISTRATE
+        .block("creative_power", props -> new MachineBlock(props, MachineBlockEntities.CREATIVE_POWER))
         .item()
         .tab(EIOCreativeTabs.MACHINES)
         .build()